extra/datasets-server/requirements.txt
fastapi==0.110.0
numpy==1.21.5
PyYAML==6.0.1
uvicorn==0.28.0
